+++ sample-tracking-status-overview-domain/src/test/groovy/life/qbic/business/samples/info/GetSamplesInfoSpec.groovy
@@ -1,7 +1,8 @@
 package life.qbic.business.samples.info
 
-import life.qbic.business.samples.download.DownloadSamplesDataSource
 import life.qbic.business.DataSourceException
+import life.qbic.business.samples.Sample
+import life.qbic.business.samples.download.DownloadSamplesDataSource
 import life.qbic.datamodel.samples.Status
 import spock.lang.Specification
 
@@ -17,39 +18,94 @@ class GetSamplesInfoSpec extends Specification {
         given:
         DownloadSamplesDataSource sampleDataSource = Stub()
         GetSamplesInfoDataSource infoDataSource = Stub()
+        SampleStatusDataSource sampleStatusDataSource = Mock()
         String projectCode = "QABCD"
-        List<String> codes = ["QABCD001AB", "QABCD002AC", "QABCD005AX", "QABCD019A2"]
+        List<String> codes = ["QABCD001A0", "QABCD002A8", "QABCD005AW", "QABCD019A3"]
+        Status expectedStatus = Status.SAMPLE_QC_FAIL
+        Sample sampleOne = new Sample("QABCD001A0", "one", expectedStatus)
+        Sample sampleTwo = new Sample("QABCD002A8", "two", expectedStatus)
+        Sample sampleGreen = new Sample("QABCD005AW", "green", expectedStatus)
+        Sample sampleBlue = new Sample("QABCD019A3", "blue", expectedStatus)
+
         Map<String, String> mapWithNames = new HashMap<>()
-        mapWithNames.put("QABCD001AB", "one")
-        mapWithNames.put("QABCD002AC", "two")
-        mapWithNames.put("QABCD005AX", "green")
-        mapWithNames.put("QABCD019A2", "blue")
-        sampleDataSource.fetchSampleCodesFor(projectCode, Status.SAMPLE_QC_FAIL) >> { codes }
+        mapWithNames.put("QABCD001A0", "one")
+        mapWithNames.put("QABCD002A8", "two")
+        mapWithNames.put("QABCD005AW", "green")
+        mapWithNames.put("QABCD019A3", "blue")
+
+        sampleDataSource.fetchSampleCodesFor(projectCode, expectedStatus) >> { codes }
         infoDataSource.fetchSampleNamesFor(codes) >> { mapWithNames }
         GetSamplesInfoOutput output = Mock()
-        GetSamplesInfo getInfos = new GetSamplesInfo(sampleDataSource, infoDataSource, output)
+        GetSamplesInfo getInfos = new GetSamplesInfo(sampleStatusDataSource, sampleDataSource, infoDataSource, output)
         
         when:"the use case is run"
-        getInfos.requestSampleInfosFor(projectCode, Status.SAMPLE_QC_FAIL)
+        getInfos.requestSampleInfosFor(projectCode, expectedStatus)
         
         then:"a successful message is send"
-        1 * output.samplesWithNames(projectCode, Status.SAMPLE_QC_FAIL, mapWithNames)
+        1 * output.samplesWithNames(*_) >> { arguments ->
+            final Collection<Sample> givenSamples = arguments[0]
+            assert givenSamples.containsAll([sampleOne, sampleTwo, sampleGreen, sampleBlue])
+        }
         0 * output.failedExecution(_ as String)
     }
 
+
+    def "successful execution of the use case forwards the map of found samples of all statuses"() {
+
+        given:
+        DownloadSamplesDataSource sampleDataSource = Stub()
+        GetSamplesInfoDataSource infoDataSource = Stub()
+        SampleStatusDataSource sampleStatusDataSource = Stub()
+        String projectCode = "QABCD"
+        List<String> codes = ["QABCD001A0", "QABCD002A8", "QABCD005AW", "QABCD019A3"]
+        Sample sampleOne = new Sample("QABCD001A0", "one", Status.DATA_AVAILABLE)
+        Sample sampleTwo = new Sample("QABCD002A8", "two", Status.DATA_AVAILABLE)
+        Sample sampleGreen = new Sample("QABCD005AW", "green", Status.SAMPLE_QC_FAIL)
+        Sample sampleBlue = new Sample("QABCD019A3", "blue", Status.SAMPLE_QC_PASS)
+
+        Map<String, String> mapWithNames = new HashMap<>()
+        mapWithNames.put("QABCD001A0", "one")
+        mapWithNames.put("QABCD002A8", "two")
+        mapWithNames.put("QABCD005AW", "green")
+        mapWithNames.put("QABCD019A3", "blue")
+
+        Map<String, Status> mapWithStatus = new HashMap<>()
+        mapWithStatus.put(sampleOne.code, sampleOne.status)
+        mapWithStatus.put(sampleTwo.code, sampleTwo.status)
+        mapWithStatus.put(sampleGreen.code, sampleGreen.status)
+        mapWithStatus.put(sampleBlue.code, sampleBlue.status)
+
+        sampleDataSource.fetchSampleCodesFor(projectCode) >> { codes }
+        infoDataSource.fetchSampleNamesFor(codes) >> { mapWithNames }
+        sampleStatusDataSource.fetchSampleStatusesFor(codes) >> { mapWithStatus }
+        GetSamplesInfoOutput output = Mock()
+        GetSamplesInfo getInfos = new GetSamplesInfo(sampleStatusDataSource, sampleDataSource, infoDataSource, output)
+
+        when:"the use case is run"
+        getInfos.requestSampleInfosFor(projectCode)
+
+        then:"a successful message is send"
+        1 * output.samplesWithNames(*_) >> { arguments ->
+            final Collection<Sample> givenSamples = arguments[0]
+            assert givenSamples.containsAll([sampleOne, sampleTwo, sampleGreen, sampleBlue])
+        }
+        0 * output.failedExecution(_ as String)
+    }
     
     def "unsuccessful execution of the use case does not lead to failure notification, but forwards the exception"() {
         
         given:
         DownloadSamplesDataSource sampleDataSource = Stub()
         GetSamplesInfoDataSource infoDataSource = Stub()
+        SampleStatusDataSource sampleStatusDataSource = Mock()
+
         String projectCode = "QABCD"
         sampleDataSource.fetchSampleCodesFor(projectCode, Status.SAMPLE_QC_FAIL) >> { new ArrayList<String>() }
         infoDataSource.fetchSampleNamesFor(_) >> { 
             throw new RuntimeException("Testing runtime exceptions")
         }
         GetSamplesInfoOutput output = Mock()
-        GetSamplesInfo getInfos = new GetSamplesInfo(sampleDataSource, infoDataSource, output)
+        GetSamplesInfo getInfos = new GetSamplesInfo(sampleStatusDataSource, sampleDataSource, infoDataSource, output)
         
         when:"the use case is run"
         getInfos.requestSampleInfosFor(projectCode, Status.SAMPLE_QC_FAIL)
@@ -66,13 +122,14 @@ class GetSamplesInfoSpec extends Specification {
         given:
         DownloadSamplesDataSource sampleDataSource = Stub()
         GetSamplesInfoDataSource infoDataSource = Stub()
+        SampleStatusDataSource sampleStatusDataSource = Mock()
         String projectCode = "QABCD"
         sampleDataSource.fetchSampleCodesFor(projectCode, Status.SAMPLE_QC_FAIL) >> { new ArrayList<String>() }
         infoDataSource.fetchSampleNamesFor(_) >> {
             throw new DataSourceException("Testing data source exception")
         }
         GetSamplesInfoOutput output = Mock()
-        GetSamplesInfo getInfos = new GetSamplesInfo(sampleDataSource, infoDataSource, output)
+        GetSamplesInfo getInfos = new GetSamplesInfo(sampleStatusDataSource, sampleDataSource, infoDataSource, output)
         
         when:"the use case is run"
         getInfos.requestSampleInfosFor(projectCode, Status.SAMPLE_QC_FAIL)